+++ src/app/(private)/user/divelogs/page.tsx
@@ -2,9 +2,7 @@
 import { Button } from "@/components/ui/button";
 import PageTitle from "@/components/ui/page-title";
 import Link from "next/link";
-import React, { useEffect } from "react";
 import { Record } from "@/interfaces";
-import { deleteRecord, getAllRecords } from "@/server-actions/record";
 import {
   Table,
   TableBody,
@@ -28,78 +26,19 @@ import { Edit2, Trash2 } from "lucide-react";
 import { useRouter } from "next/navigation";
 import Spinner from "@/components/ui/spinner";
 import InfoMessages from "@/components/ui/info-messages";
-import toast from "react-hot-toast";
-import { User } from "@supabase/supabase-js";
-import { getLoggedInUser } from "@/server-actions/users";
 import { truncateHtml } from "@/helpers/truncate-html";
+import { useDivelogs } from "@/hooks/useDivelogs";
 
 function userDivelogPage() {
   const router = useRouter();
-  const [loading, setLoading] = React.useState(true);
-  const [records, setRecords] = React.useState<Record[]>([]);
-  const [user, setUser] = React.useState<User | null>(null);
-  const [deleteDialogOpen, setDeleteDialogOpen] = React.useState(false);
-  const [recordToDelete, setRecordToDelete] = React.useState<number | null>(
-    null
-  );
-
-  const fetchUser = async () => {
-    const response = await getLoggedInUser();
-    if (response.success) {
-      setUser(response.data);
-    } else {
-      if (response.message !== undefined) {
-        toast.error(response.message);
-      }
-    }
-  };
-
-  useEffect(() => {
-    fetchUser();
-  }, []);
-
-  const fetchRecords = async (user_id: number) => {
-    setLoading(true);
-    const response = await getAllRecords(user_id);
-    if (response.success) {
-      setRecords(response.data);
-    } else {
-      setRecords([]);
-    }
-    setLoading(false);
-  };
-
-  const openDeleteDialog = (id: number) => {
-    setRecordToDelete(id);
-    setDeleteDialogOpen(true);
-  };
-
-  const handleDelete = async () => {
-    if (recordToDelete === null) return;
-
-    setLoading(true);
-    setDeleteDialogOpen(false);
-
-    const response = await deleteRecord(recordToDelete);
-    if (response.success) {
-      setRecords((prev) => prev.filter((rcd) => rcd.id !== recordToDelete));
-      if (response.message !== undefined) {
-        toast.success(response.message);
-      }
-    } else {
-      if (response.message !== undefined) {
-        toast.error(response.message);
-      }
-    }
-    setLoading(false);
-    setRecordToDelete(null);
-  };
-
-  useEffect(() => {
-    if (user?.id) {
-      fetchRecords(Number(user.id));
-    }
-  }, [user?.id]);
+  const {
+    loading,
+    records,
+    deleteDialogOpen,
+    setDeleteDialogOpen,
+    openDeleteDialog,
+    handleDelete,
+  } = useDivelogs();
 
   const columns = [
     "Rate",
